+++ src/main/java/net/sf/jabref/logic/integrity/EntryLinkChecker.java
@@ -11,7 +11,7 @@ import net.sf.jabref.logic.integrity.IntegrityCheck.Checker;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.model.entry.BibEntry;
-import net.sf.jabref.model.entry.FieldProperties;
+import net.sf.jabref.model.entry.FieldProperty;
 import net.sf.jabref.model.entry.InternalBibtexFields;
 
 
@@ -28,18 +28,19 @@ public class EntryLinkChecker implements Checker {
     public List<IntegrityMessage> check(BibEntry entry) {
         List<IntegrityMessage> result = new ArrayList<>();
         for (Entry<String,String> field : entry.getFieldMap().entrySet()) {
-            Set<FieldProperties> properties = InternalBibtexFields.getFieldExtras(field.getKey());
-            if (properties.contains(FieldProperties.SINGLE_ENTRY_LINK)) {
+            Set<FieldProperty> properties = InternalBibtexFields.getFieldProperties(field.getKey());
+            if (properties.contains(FieldProperty.SINGLE_ENTRY_LINK)) {
                 if (!database.getEntryByKey(field.getValue()).isPresent()) {
-                    result.add(new IntegrityMessage(Localization.lang("BibTeX key does not exist"), entry,
+                    result.add(new IntegrityMessage(Localization.lang("Referenced BibTeX key does not exist"), entry,
                             field.getKey()));
                 }
-            } else if (properties.contains(FieldProperties.MULTIPLE_ENTRY_LINK)) {
+            } else if (properties.contains(FieldProperty.MULTIPLE_ENTRY_LINK)) {
                 List<String> keys = new ArrayList<>(Arrays.asList(field.getValue().split(",")));
                 for (String key : keys) {
                     if (!database.getEntryByKey(key).isPresent()) {
-                        result.add(new IntegrityMessage(Localization.lang("BibTeX key does not exist") + ": " + key,
-                                entry, field.getKey()));
+                        result.add(new IntegrityMessage(
+                                Localization.lang("Referenced BibTeX key does not exist") + ": " + key, entry,
+                                field.getKey()));
                     }
                 }
             }